[FIX] hr_timesheet_ux: Fix stat buttons of timesheet sheet form
This one was not working, now they do. We was not able to use the already
exist actions because we need to filter by current timesheet sheet employee
instead of active user. Also need to filter by the timesheet sheet dates
instead of current week as it was by default in the attendee vs timesheet
report.