[IMP] hr_timesheet_ux: Multiple improves (#12)
* [ADD] hr_timesheet_ux: Hide My Timesheets/All Timesheets menus. set then
visible in developer mode.

* [FIX] hr_timesheet_ux: remove old translations of hr_timesheet_balance module.

* [FIX] hr_timesheet_ux: remove execution permissions for view file.

* [FIX] hr_timesheet_ux: Fix stat buttons of timesheet sheet form

This one was not working, now they do. We was not able to use the already
exist actions because we need to filter by current timesheet sheet employee
instead of active user. Also need to filter by the timesheet sheet dates
instead of current week as it was by default in the attendee vs timesheet
report.

* [FIX] hr_timesheet_ux: Compute attendence

Also add functionality to compute attendense without need to set a
check_out date.